Make app_author a required parameter
Windows requires app_author to have a meaningful value, while other
OSs don't use it. To make a unified API, we make app_author a
required parameter on all OSs. It is now ignored in linux and macOS
and can be anything. On Windows it must be a nonempty string.